+++ aws_terraform_registry/common/__init__.py
@@ -1,6 +1,6 @@
 from .logger import init_root_logger
 from .model import TerraformModuleIdentifier
-from .publish import publish_module
+from .publish import publish_module, unpublish_module
 from .release import release_module
 from .token import generate_terraformrc, generate_token
 
@@ -10,5 +10,6 @@ __all__ = [
     'generate_terraformrc',
     'TerraformModuleIdentifier',
     'publish_module',
+    'unpublish_module',
     'release_module',
 ]