+++ aws_terraform_registry/common/release.py
@@ -6,8 +6,8 @@ from pathlib import Path
 from boto3 import client
 
 from ..config import ApplicationConfig
-from .model import TerraformModuleIdentifier
-from .publish import publish_module
+from .model import BUCKET_FILE_NAME, TerraformModuleIdentifier
+from .publish import exists, publish_module
 
 logger = getLogger()
 
@@ -35,11 +35,16 @@ def release_module(
         source (str): module source
 
     Raise:
-        (RuntimeError): if source did not exists
+        (RuntimeError): if source did not exists or if specified version and module is ever published.
     """
     config.validate()
     # remove the v
     version = version if not version.lower().startswith("v") else version[1:]
+
+    if exists(config=config, terraform_module=terraform_module, version=version):
+        msg = f"Version {version} for module {terraform_module.module_id} ever exist"
+        logger.error(msg)
+        raise RuntimeError(msg)
 
     s3_key = terraform_module.get_bucket_key(version=version)
     logger.debug(f"Put module archive to {s3_key}")
@@ -56,6 +61,8 @@ def release_module(
         send_s3_from_dir(config=config, archive_dir=_source, s3_key=s3_key)
 
     publish_url = terraform_module.get_publish_url(bucket_name=config.bucket_name, version=version)
+    # experimental API
+    # publish_url = terraform_module.get_blob_url(repository_url=config.repository_url, version=version)
     logger.debug(f"url: {publish_url}")
     publish_module(config=config, terraform_module=terraform_module, version=version, source=publish_url)
 
@@ -69,7 +76,7 @@ def send_s3_from_file(config: ApplicationConfig, archive_file: str, s3_key: str)
 
 
 def send_s3_from_dir(config: ApplicationConfig, archive_dir: str, s3_key: str):
-    archive_file = Path.cwd() / "archive.tar.gz"
+    archive_file = Path.cwd() / BUCKET_FILE_NAME
     try:
         with tarfile.open(archive_file, "w:gz") as tar:
             tar.add(archive_dir, arcname=".", filter=lambda a: a if not a.name.startswith("./.") else None)
@@ -80,7 +87,7 @@ def send_s3_from_dir(config: ApplicationConfig, archive_dir: str, s3_key: str):
 
 def send_s3_from_url(config: ApplicationConfig, source_url: str, s3_key: str):
     opener = urllib.request.build_opener()
-    archive_file = Path.cwd() / "archive.tar.gz"
+    archive_file = Path.cwd() / BUCKET_FILE_NAME
     try:
         with open(archive_file, "wb") as archive:
             with opener.open(source_url) as object_data: